// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/wearable-connection-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface WearableConnectionHttpResponse {
    connectedAtUtc?: string;
    externalUserId?: string;
    isActive?: boolean;
    lastSyncedAtUtc?: string | null;
    provider?: string;
}