// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/wearable-daily-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface WearableDailySummaryHttpResponse {
    activeMinutes?: number | null;
    caloriesBurned?: number | null;
    date?: string;
    heartRate?: number | null;
    sleepMinutes?: number | null;
    steps?: number | null;
}